+++ packages/core/analytics-next/src/__tests__/unit/server-side-hydrate.js
@@ -13,8 +13,8 @@ beforeAll(() => {
 afterEach(() => {
   jest.resetAllMocks();
 });
-
-test('should ssr then hydrate analytics-next correctly', async () => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('should ssr then hydrate analytics-next correctly', async () => {
   const [example] = await getExamplesFor('analytics-next');
   // $StringLitteral
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require