+++ packages/core/feedback-collector/src/__tests__/unit/server-side-hydrate.js
@@ -13,8 +13,8 @@ beforeEach(() => {
 afterEach(() => {
   jest.resetAllMocks();
 });
-
-test('should ssr then hydrate feedback-collector correctly', async () => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('should ssr then hydrate feedback-collector correctly', async () => {
   const [example] = await getExamplesFor('feedback-collector');
   // $StringLitteral
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require