+++ packages/core/input/src/SingleLineTextInput.js
@@ -55,6 +55,13 @@ type Props = {
   onKeyDown?: (e: KeyboardEvent) => mixed,
 };
 
+if (process.env.NODE_ENV !== 'production') {
+  // eslint-disable-next-line no-console
+  console.warn(
+    '@atlaskit/input has been deprecated. It is an internal component and should not be used directly.',
+  );
+}
+
 export default class SingleLineTextInput extends Component<Props, {}> {
   static defaultProps = {
     style: {},