+++ packages/core/radio/src/__tests__/unit/server-side-hydrate.tsx
@@ -9,10 +9,10 @@ jest.spyOn(global.console, 'error').mockImplementation(() => {});
 afterEach(() => {
   jest.resetAllMocks();
 });
-
-test('should ssr then hydrate radio correctly', async () => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('should ssr then hydrate radio correctly', async () => {
   const [example] = await getExamplesFor('radio');
-  // $StringLitteral
+
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
 
   const elem = document.createElement('div');