+++ packages/editor/editor-core/src/__tests__/visual-regression/media/lists.ts
@@ -1,4 +1,4 @@
-import { snapshot, initFullPageEditorWithAdf, Device } from '../_utils';
+import { snapshot, initEditorWithAdf, Appearance } from '../_utils';
 import {
   clickEditableContent,
   typeInEditor,
@@ -8,52 +8,56 @@ import {
   waitForMediaToBeLoaded,
   resizeMediaInPosition,
   scrollToMedia,
+  clickMediaInPosition,
 } from '../../__helpers/page-objects/_media';
 import * as panelList from './__fixtures__/panel-list-adf.json';
 import { Page } from '../../__helpers/page-objects/_types';
 
+let page: Page;
+
+const initEditor = async (adf?: Object) => {
+  await initEditorWithAdf(page, {
+    appearance: Appearance.fullPage,
+    adf,
+    viewport: { width: 1040, height: 750 },
+  });
+  await clickEditableContent(page);
+};
+
 describe('Snapshot Test: Media', () => {
-  let page: Page;
   beforeEach(async () => {
     // @ts-ignore
     page = global.page;
   });
 
-  describe('Lists', async () => {
+  describe('Lists', () => {
     beforeEach(async () => {
-      await initFullPageEditorWithAdf(page, {}, Device.LaptopHiDPI);
-      await clickEditableContent(page);
+      await initEditor();
+    });
+
+    afterEach(async () => {
+      await insertMedia(page);
+      await scrollToMedia(page);
+      await waitForMediaToBeLoaded(page);
+      await clickMediaInPosition(page, 0); // want to see resize handles
+
+      await snapshot(page);
     });
 
     it('can insert a media single inside a bullet list', async () => {
       await typeInEditor(page, '* ');
-
-      // now we can insert media as necessary
-      await insertMedia(page);
-      await scrollToMedia(page);
-      await waitForMediaToBeLoaded(page);
-
-      await snapshot(page);
     });
 
     it('can insert a media single inside a numbered list', async () => {
       // type some text
       await typeInEditor(page, '1. ');
-
-      // now we can insert media as necessary
-      await insertMedia(page);
-      await scrollToMedia(page);
-      await waitForMediaToBeLoaded(page);
-
-      await snapshot(page);
     });
   });
 
   // TODO: Convert to integration test (https://product-fabric.atlassian.net/browse/ED-6692)
-  describe('Lists in panels', async () => {
+  describe('Lists in panels', () => {
     beforeEach(async () => {
-      await initFullPageEditorWithAdf(page, panelList, Device.LaptopHiDPI);
-      await clickEditableContent(page);
+      await initEditor(panelList);
     });
 
     it('can be resized in a list in a panel', async () => {