+++ packages/editor/editor-core/src/plugins/table/commands-with-analytics.ts
@@ -25,6 +25,7 @@ import {
   deleteTable,
   toggleTableLayout,
   getNextLayout,
+  sortByColumn,
 } from './commands';
 import {
   getSelectedCellInfo,
@@ -34,6 +35,7 @@ import {
 import { getPluginState } from './pm-plugins/main';
 import { mergeCells, deleteColumns, deleteRows } from './transforms';
 import { splitCell } from './commands/misc';
+import { SortOrder } from './types';
 
 const TABLE_BREAKOUT_NAME_MAPPING = {
   default: TABLE_BREAKOUT.NORMAL,
@@ -406,4 +408,27 @@ export const toggleTableLayoutWithAnalytics = () =>
     }
     return;
   })(toggleTableLayout);
+
+export const sortColumnWithAnalytics = (
+  inputMethod: INPUT_METHOD.CONTEXT_MENU,
+  columnIndex: number,
+  sortOrder: SortOrder,
+) =>
+  withAnalytics(state => {
+    const { totalRowCount, totalColumnCount } = getSelectedTableInfo(
+      state.selection,
+    );
+    return {
+      action: TABLE_ACTION.SORTED_COLUMN,
+      actionSubject: ACTION_SUBJECT.TABLE,
+      attributes: {
+        inputMethod,
+        totalRowCount,
+        totalColumnCount,
+        position: columnIndex,
+        sortOrder,
+      },
+      eventType: EVENT_TYPE.TRACK,
+    };
+  })(sortByColumn(columnIndex, sortOrder));
 // #endregion