+++ packages/media/media-picker/src/components/dropzone/dropzone.tsx
@@ -1,3 +1,4 @@
+import * as React from 'react';
 import {
   LocalUploadComponentReact,
   LocalUploadComponentBaseProps,
@@ -10,25 +11,39 @@ import {
   DropzoneUploadEventPayloadMap,
 } from '../types';
 
-const toArray = (arr: any) => [].slice.call(arr, 0);
+import {
+  withAnalyticsEvents,
+  withAnalyticsContext,
+  WithAnalyticsEventProps,
+  AnalyticsEventPayload,
+} from '@atlaskit/analytics-next';
 
-export type DropzoneProps = LocalUploadComponentBaseProps & {
-  config: DropzoneConfig;
-  onDrop?: () => void;
-  onDragEnter?: (payload: DropzoneDragEnterEventPayload) => void;
-  onDragLeave?: (payload: DropzoneDragLeaveEventPayload) => void;
-  onCancelFn?: (cancel: (uploadId: string) => void) => void;
-};
+import { FabricChannel } from '@atlaskit/analytics-listeners';
+
+import {
+  name as packageName,
+  version as packageVersion,
+} from '../../version.json';
+
+export type DropzoneProps = LocalUploadComponentBaseProps &
+  WithAnalyticsEventProps &
+  React.RefAttributes<DropzoneBase> & {
+    config: DropzoneConfig;
+    onDrop?: () => void;
+    onDragEnter?: (payload: DropzoneDragEnterEventPayload) => void;
+    onDragLeave?: (payload: DropzoneDragLeaveEventPayload) => void;
+    onCancelFn?: (cancel: (uploadId: string) => void) => void;
+  };
 
 function dragContainsFiles(event: DragEvent): boolean {
   if (!event.dataTransfer) {
     return false;
   }
   const { types } = event.dataTransfer;
-  return toArray(types).indexOf('Files') > -1;
+  return Array.from(types).indexOf('Files') > -1;
 }
 
-export class Dropzone extends LocalUploadComponentReact<
+export class DropzoneBase extends LocalUploadComponentReact<
   DropzoneProps,
   DropzoneUploadEventPayloadMap
 > {
@@ -127,8 +142,9 @@ export class Dropzone extends LocalUploadComponentReact<
     dragEvent.stopPropagation();
     this.onDrop(dragEvent);
 
-    const filesArray = [].slice.call(dragEvent.dataTransfer.files);
-    this.uploadService.addFiles(filesArray);
+    const files = Array.from(dragEvent.dataTransfer.files);
+
+    this.uploadService.addFiles(files);
   };
 
   // Cross-browser way of getting dragged items length, we prioritize "items" if present
@@ -138,7 +154,7 @@ export class Dropzone extends LocalUploadComponentReact<
   // available on 'drop'
   private getDraggedItemsLength(dataTransfer: DataTransfer): number {
     if (dataTransfer.items) {
-      const items = toArray(dataTransfer.items);
+      const items = Array.from(dataTransfer.items);
       return items.filter((i: DataTransferItem) => i.kind === 'file').length;
     }
     // This is required for IE11
@@ -149,16 +165,34 @@ export class Dropzone extends LocalUploadComponentReact<
     if (e.dataTransfer && dragContainsFiles(e)) {
       const dataTransfer = e.dataTransfer;
       const length = this.getDraggedItemsLength(dataTransfer);
+
+      this.fireAnalyticsEvent({
+        action: 'droppedInto',
+        actionSubject: 'dropzone',
+        attributes: {
+          fileCount: length,
+        },
+      });
+
       if (this.props.onDrop) this.props.onDrop();
       this.emitDragLeave({ length });
     }
   };
 
-  private emitDragOver(e: DropzoneDragEnterEventPayload): void {
+  private emitDragOver(payload: DropzoneDragEnterEventPayload): void {
     if (!this.uiActive) {
       const { onDragEnter } = this.props;
       this.uiActive = true;
-      if (onDragEnter) onDragEnter(e);
+
+      this.fireAnalyticsEvent({
+        action: 'draggedInto',
+        actionSubject: 'dropzone',
+        attributes: {
+          fileCount: payload.length,
+        },
+      });
+
+      if (onDragEnter) onDragEnter(payload);
     }
   }
 
@@ -172,9 +206,26 @@ export class Dropzone extends LocalUploadComponentReact<
       window.setTimeout(() => {
         if (!this.uiActive) {
           const { onDragLeave } = this.props;
+
+          this.fireAnalyticsEvent({
+            action: 'draggedOut',
+            actionSubject: 'dropzone',
+            attributes: {
+              fileCount: payload.length,
+            },
+          });
+
           if (onDragLeave) onDragLeave(payload);
         }
       }, 50);
+    }
+  }
+
+  private fireAnalyticsEvent(payload: AnalyticsEventPayload): void {
+    const { createAnalyticsEvent } = this.props;
+    if (createAnalyticsEvent) {
+      const analyticsEvent = createAnalyticsEvent(payload);
+      analyticsEvent.fire(FabricChannel.media);
     }
   }
 
@@ -182,3 +233,9 @@ export class Dropzone extends LocalUploadComponentReact<
     return null;
   }
 }
+
+export const Dropzone = withAnalyticsContext<DropzoneProps>({
+  componentName: 'dropzone',
+  packageName,
+  packageVersion,
+})(withAnalyticsEvents<DropzoneProps>()(DropzoneBase));