+++ packages/media/media-viewer/src/__tests__/unit/newgen/media-viewer-server-side.tsx
@@ -6,8 +6,8 @@ import waitForExpect from 'wait-for-expect';
 beforeAll(() => {
   jest.setTimeout(20000);
 });
-
-test('media-viewer server side rendering', async () => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('media-viewer server side rendering', async () => {
   const examples = await getExamplesFor('media-viewer');
   for (const example of examples) {
     const Example = await require(example.filePath).default;