+++ packages/core/navigation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/navigation
 
+## 35.1.14
+
+### Patch Changes
+
+- [patch][226a5fece8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/226a5fece8):
+
+  Upating deprecation messages and adding console warning to improve visibility
+
 ## 35.1.13
 
 ### Patch Changes